Turned off Discovery address by default, added security guards ===============================================================
System Information:
Engine Version: Molecule Ver 2.6.7.20250609
Library Version: Ver 3.4.1.20250521
ABI: Ver 1.5.0.20250521
Common: Ver 3.3.0.20250521

Hosts:
Interactive Console Ver 1.0.8.20250609
Pipe Child Ver 1.1.5.20250609
Headless Host Ver 1.1.1.20250609

Storages:
File System Storage Ver 1.4.1.20250609

Firmware:
Windows Ver 2.3.0.20250520
Linux Ver 2.1.1.20250522

Tested Compilers:
MSVC 14.44.35207
GCC 11.4.0

Tested OSes:
Linux Ver:6.8.0-60-generic [x86_64]
Windows Workstation 10.0.26100.4061 [x64]

+++ common/rx_value_conv.c
@@ -2086,7 +2086,7 @@ RX_COMMON_API int rx_convert_value(struct typed_value_type* val, rx_value_t type
 			if (ret == RX_OK)
 			{
 				// everything is OK so destroy previous and assign new ones
-				for (int j = 0; j < val->value.array_value.size; j++)
+				for (size_t j = 0; j < val->value.array_value.size; j++)
 				{
 					destroy_union_value(&val->value.array_value.values[j], val->value_type & RX_STRIP_ARRAY_MASK);
 					val->value.array_value.values[j] = temp[j];
@@ -2095,7 +2095,7 @@ RX_COMMON_API int rx_convert_value(struct typed_value_type* val, rx_value_t type
 			}
 			else if(i > 0)
 			{
-				for (int j = 0; j < i; j++)
+				for (size_t j = 0; j < i; j++)
 				{
 					destroy_union_value(&temp[j], type & RX_STRIP_ARRAY_MASK);
 				}
